[DF-185] fixed warnings
Former-commit-id: 8e33d1517964a2d8b88287848aa7834670d654b4

+++ Tools/ResourceEditor/Classes/Qt/Scene/System/LandscapeEditorDrawSystem/VisibilityToolProxy.cpp
@@ -77,7 +77,7 @@ Rect VisibilityToolProxy::GetChangedRect()
 
 void VisibilityToolProxy::UpdateRect(const DAVA::Rect &rect)
 {
-	Rect bounds(0.f, 0.f, size - 1, size - 1);
+	Rect bounds(0.f, 0.f, (float32)(size - 1), (float32)(size - 1));
 	changedRect = rect;
 	bounds.ClampToRect(changedRect);
